vidhyaSetu/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.63",
    "fastapi>=0.135.1",
    "groq>=1.0.0",
    "mangum>=0.21.0",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
